cleanup crate structure (#450)
