+++ git-refspec/tests/parse/fetch.rs
@@ -1,5 +1,5 @@
 use crate::parse::{assert_parse, b, try_parse};
-use git_refspec::{parse::Error, Fetch, Instruction, Operation};
+use git_refspec::{instruction::Fetch, parse::Error, parse::Operation, Instruction};
 
 #[test]
 fn revspecs_are_disallowed() {
